backend/requirements.txt
<<<<<<< HEAD
# Core app (app/) uses only the Python standard library — no extra packages
# to import or run algorithms/simulation. Use Python 3.9+.
#
# For running tests from the `backend` folder:
#   python -m pip install -r requirements.txt
#   python -m pytest -v

pytest>=7.0
3
=======
fastapi
uvicorn
redis
upstash-redis
pydantic
pydantic-settings
websockets
python-dotenv
numpy
asyncio
>>>>>>> 894b085d3ca0d253b68a3a4263d9aee0bb08665b
